move "unseenMessages" reducer into own file - reuse existing actions for tracking and resetting unseen messages - add "ui middleware" to dispatch "setOpen" as a sideeffect of "toggleOpen" - add any type of message to unseen regardless of content
