+++ src/webchat/store/connection/connection-middleware.ts
@@ -4,7 +4,6 @@ import { SetOpenAction, ToggleOpenAction } from "../ui/ui-reducer";
 import { SendMessageAction } from "../messages/message-middleware";
 import { setOptions } from "../options/options-reducer";
 import { SocketClient } from "@cognigy/socket-client";
-import { addMessageTeaser, clearUnseenMessages, removeMessageTeaser, setMessageTeaser } from "../messages/message-reducer";
 
 export interface ISendMessageOptions {
     /* overrides the displayed text within a chat bubble. useful for e.g. buttons */
@@ -45,10 +44,6 @@ export const createConnectionMiddleware = (client: SocketClient): Middleware<{},
             if (!store.getState().ui.open) {
                 store.dispatch(connect())
             }
-
-            // Remove all unseen messages
-            store.dispatch(clearUnseenMessages());
-            break;
         }
 
         case 'SEND_MESSAGE': {
