+++ src/webchat/store/store.ts
@@ -11,6 +11,7 @@ import { createAnalyticsMiddleware } from './analytics/analytics-middleware';
 import { registerConnectionHandler } from './connection/connection-handler';
 import { Webchat } from '../components/Webchat';
 import { IWebchatSettings } from '../../common/interfaces/webchat-config';
+import { uiMiddleware } from './ui/ui-middleware';
 
 
 export type StoreState = StateType<typeof reducer>;
@@ -27,6 +28,7 @@ export const createWebchatStore = (webchat: Webchat, url: string, overrideWebcha
             createMessageMiddleware(client),
             createConfigMiddleware(url, overrideWebchatSettings),
             optionsMiddleware,
+            uiMiddleware
         )
     );
